Tarea 1/server/requirements.txt
psycopg2-binary
redis
grpcio
grpcio-tools
protobuf
